@@ -29,10 +29,10 @@
     "@types/node": "^20.0.0",
     "@types/react": "18.3.1",
     "@types/react-dom": "18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.16.0",
-    "@typescript-eslint/parser": "^7.16.0",
+    "@typescript-eslint/eslint-plugin": "~8.49.0",
+    "@typescript-eslint/parser": "~8.49.0",
     "babel-jest": "^29.7.0",
-    "eslint": "~8.57.0",
+    "eslint": "~9.39.2",
     "eslint-config-next": "15.2.4",
     "eslint-config-prettier": "10.0.0",
     "eslint-plugin-import": "2.31.0",
